requirements.txt
# Gmail API and Authentication
google-api-python-client==2.100.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Natural Language Processing
nltk==3.8.1

# Task Scheduling
schedule==1.2.0

# HTTP Requests
httplib2==0.22.0


